+++ API-Ferreteria/src/controllers/balances.controller.js
@@ -1,1 +1,7 @@
 import { Connection, getConnection } from "./../database/database"
+
+
+
+export const balancesMetodos = {
+    
+}
