+++ API-Ferreteria/src/controllers/transacciones.controller.js
@@ -1,1 +1,5 @@
 import { Connection, getConnection } from "./../database/database"
+
+export const transaccionesMetodos = {
+    
+}
